Refact/create filesContext (#90)
* refact: add filesContext
