+++ frontend/src/components/annotation/AnnotationImage.tsx
@@ -2,11 +2,11 @@ import { useTranslation } from "react-i18next";
 import { capitalize, Grid } from "@mui/material";
 import AnnotationImageDisplay from "./AnnotationImageDisplay";
 import AnnotationImageNavigation from "./AnnotationImageNavigation";
-import { useMainContext } from "../../contexts/mainContext";
+import { useFilesContext } from "../../contexts/filesContext";
 
 export default function AnnotationImage() {
   const { t } = useTranslation();
-  const { image } = useMainContext();
+  const { image } = useFilesContext();
 
   return (
     <>
