+++ frontend/src/components/annotation/AnnotationImageNavigation.tsx
@@ -1,5 +1,5 @@
 import { Grid, IconButton, Typography } from "@mui/material";
-import { useMainContext } from "../../contexts/mainContext";
+import { useFilesContext } from "../../contexts/filesContext";
 import { useAnnotationContext } from "../../contexts/annotationContext";
 import SkipNextIcon from '@mui/icons-material/SkipNext';
 import SkipPreviousIcon from '@mui/icons-material/SkipPrevious';
@@ -9,7 +9,7 @@ import FastRewindIcon from '@mui/icons-material/FastRewind';
 
 const AnnotationImageNavigation = () => {
 
-    const { files, currentImage } = useMainContext();
+    const { files, currentImage } = useFilesContext();
     const { previous, lastOrFirstImage, next } = useAnnotationContext();
 
     const imageIndex = () => {
